requirements.txt
pydub >= 0.25.1
pysimplegui >= 4.60.1
pytube >= 12.1.0
ffmpeg-python >= 0.2.0